crypto: split up WIP swap code into modules
This reorganization divides the existing swap.rs into submodules, and puts it
under a new `dex` module, so that we can have a place for structures that are
common to both the swap side and the lp side, and have room for sibling `swap`
and `lp` modules.
