+++ crypto/src/dex.rs
@@ -1,0 +1,4 @@
+pub mod swap;
+
+mod trading_pair;
+pub use trading_pair::TradingPair;
