+++ crypto/src/lib.rs
@@ -8,6 +8,7 @@ pub use decaf377_rdsa as rdsa;
 mod address;
 pub mod asset;
 mod delegation_token;
+pub mod dex;
 pub mod eddy;
 mod flow;
 mod identity_key;
@@ -18,7 +19,6 @@ mod note_payload;
 mod nullifier;
 mod prf;
 pub mod proofs;
-pub mod swap;
 pub mod transaction;
 pub mod value;
